+++ lib/flutter_linkable_text.dart
@@ -1,3 +1,0 @@
-library flutter_linkable_text;
-
-export 'src/text_funtions.dart';