+++ lib/flutter_string_utils.dart
@@ -1,0 +1,5 @@
+library flutter_linkable_text;
+
+export 'src/flt_date_extensions.dart';
+export 'src/flt_string_extensions.dart';
+export 'src/text_funtions.dart';