+++ lib/src/flt_date_extensions.dart
@@ -1,0 +1,8 @@
+import 'package:intl/intl.dart';
+
+extension FLTDateTimeExtensions on DateTime {
+  String get toDateStringWithFormatDDMMYYYY {
+    DateFormat dateFormat = DateFormat('dd/MM/yyyy');
+    return dateFormat.format(this);
+  }
+}